test(ai): consolidate observer NW-suppression predicate pins into matrix (Refs #3749)
Folds the two per-predicate observer-phase NW-suppression
`*_branches_test.dart` suites (`shouldSuppressNewWorldColonialOrders`,
`shouldSuppressNewWorldDeclareWarInvasionAndPurchase`) into a single
table-driven `observer_goal_phase_nw_suppression_predicate_game_matrix_test.dart`,
mirroring the existing EXPAND `(game,snapshot)->bool` matrix consolidation
from #3809.

Both predicates share the exact `({snapshot, game}) -> bool` signature and the
same four phase fixtures (EXPAND, COLONIAL-lite, COLONIAL, DEVELOP), so each
former phase-branch test becomes one matrix row that re-asserts the shared
phase fixture control (`observerGoalPhaseFor`) and the predicate's verbatim
expected boolean + regression reason. Coverage is preserved 1:1 (10 tests in,
10 tests out); both determinism guards are retained verbatim.

Advances AC7 (expand/observer branch-pin consolidation). Remaining AC7 work
(the `({game,snapshot})->List<String>` observer peace-target deciders and the
`stalledBelowQuotaGpLeadPeaceTargets` DiplomacyRelation-backed decider) is left
to follow-up commits on this PR.

Refs #3749
